API

IonSim.Chamber — Type

Chamber(;
        iontrap::LinearChain, B::Real=0, Bhat::NamedTuple{(:x,:y,:z)=ẑ, ∇B::Real=0,
         δB::Union{Real,Function}=0, lasers::Vector{Laser}
)

Information necessary to describe the Hamiltonian for a collection of ions in a linear chain interacting with laser light. user-defined fields

iontrap<:LinearChain

B: A real value describing the mean magnitude of the B-field [Tesla].

Bhat::NamedTuple{(:x,:y,:z)}: Describes the direction of the B-field (defaults to ẑ).

∇B: Magnitude of the B-field gradient. We assume that the gradient always points along the z-direction. [Tesla / meter]

δB::Function: Time-dependence of the B-field [Tesla]

lasers::Array{<:Laser}: For each laser in the array, the pointing field should contain an array of Tuple{Int,Real}. The first element specifies the index of an ion in the ions field that the laser interacts with. The second element specifies a scaling factor for the strength of that interaction (to be used, e.g., for modeling cross-talk).

derived fields

_cnst_δB::Bool: A Boolean flag signifying whether or not δB is a constant function.

basis<:CompositeBasis: The basis for describing the combined system, ions + vibrational modes. If constructing the Hamiltonian explictly (with hamiltonian), then the ordering of the basis is set, by convention, as $ion₁ ⊗ ion₂ ⊗ ... ⊗ ion_N ⊗ mode₁ ⊗ mode₂ ⊗ ... ⊗ mode_N$, where the ion bases are ordered according to the order in T.iontrap.ions and the vibrational modes are ordered according to the order in [T.iontrap.selectedmodes.x, T.iontrap.selectedmodes.y, T.iontrap.selectedmodes.z]. E.g. for:

chain = LinearChain(ions=[C1, C2], comfrequencies=(x=2e6,y=2e6,z=1e6), selectedmodes=(x=[1, 2], y=[], z=[1]))

The ordering of the basis would be

C1.basis ⊗ C2.basis ⊗ chain.selectedmodes.x[1].basis ⊗ chain.selectedmodes.x[2].basis ⊗ chain.selectedmodes.z[1].basis

Otherwise, the ordering is according to the form of the initial state used in the solver.

IonSim.Ion — Type

Ion

The physical parameters defining an isolated ion's internal structure.

IonSim.IonInstance — Type

IonInstance(selected_sublevels::Vector{Tuple}[, manualshift::Dict]) Ion instance of some species

selected_sublevels specifies which energy sublevels will be present in the Hilbert space of this Ion instance, as a subset of all possible sublevels.

Each element of selected_sublevels is a 2-element Tuple (level::String, sublevels), with the first element being the name of a level and the second specifying which sublevels should be included. Allowed sublevels are those whose magnetic quantum number m is in the set {-f, -f+1, -f+2, ... f-1, f}, where f is the total angular momentum quantum number of level. For each level specified there are three allowed options to specify the set of sublevels to include:

sublevels::Real: Includes only one m = sublevels

sublevels::Vector{Real}: Includes all sublevels whose magnetic quantum number m is in sublevels

sublevels = "all": Includes all allowed sublevels

If an element of selected_sublevels utilizes the first option, specifying a single m, one may optionally may this a 3-element tuple instead: (level::String, m::Real, alias::String), assinging this particular sublevel the alias alias.

Omission of a level in selected_sublevels will exclude all sublevels.

Fields

speciesproperties::IonProperties: Contains constants specifying parameters specific to species

sublevels::Vector{Tuple{String,Real}}: List of all sublevels present in the Hilbert space

sublevelaliases::Dict{String,Tuple}: Dict specifying aliases assigned to sublevels, in the format alias => sublevel

shape::Vector{Int}: Dimension of the Hilbert space

manualshift::OrderedDict: A dictionary with keys denoting the selected levels and values, a real number for describing a shift of the level's energy. This is just a convenient way to add manual shifts to the simulation, such as Stark shifts off of energy levels not present in the Hilbert space, without additional resources

ionnumber: When the ion is added to an IonTrap, this value keeps track of its order

ionposition: When the ion is added to an IonTrap, this value keeps track of its physical position in meters

IonSim.IonProperties — Type

IonProperties type.

Required keywords

mass: Mass of ion in kg

charge: Charge of ion in units of elementary charge

full_level_structure: OrderedDict describing properties of each energy level

key::String: Name of energy level. Spectroscopic notation encouraged, e.g. "S1/2,f=1"

value::NamedTuple(:n, :l, :j, :f, :E): Quantum numbers n, l, j, f, and energy E (in Hz)

full_transitions: Dict of all allowed transitions between energy levels

key::Tuple{String,String} Pair of levels, ordered (lower, upper) in energy

value::NamedTuple(:multipole, :einsteinA): Leading-order multipole of the transition (e.g. "E1", "E2") and Einstein A coefficient (between fine structure levels only; hyperfine factors are calculated when needed)

Optional keywords

default_sublevel_selection: Default value of selected_sublevels argument in Ion constructor

gfactors: Dict(level::String => g::Real) Custom Landé g-factors, if contributions from higher-than-first-order perturbations are desired

nonlinear_zeeman: Dict describing nonlinear contributions to Zeeman shift of certain sublevels

key::Tuple{String,Real}: sublevel name

value::Function(B::Real): Nonlinear term(s) of Zeeman shift. Full Zeeman shift will be calculated as the sum of the usual linear term and this function

IonSim.IonSimBasis — Type

IonSimBasis

An abstract type for specialized bases, which are unique to IonSim.

IonSim.IonTrap — Type

IonTrap

A physical configuration of ions. Stores a collection of ions and information about their relative center of mass motion.

IonSim.Laser — Type

Laser(;λ=missing, E=0, Δ=0, ϵ=(x̂+ŷ)/√2, k=ẑ, ϕ=0, pointing::Array{Tuple{Int,Real}})

The physical parameters defining laser light. args

λ::Union{Real,Missing}: the wavelength of the laser in meters

I::Union{Function,Real}: laser intensity in W/m²

Δ: static detuning from f = c/λ in [Hz]

ϵ::NamedTuple: (ϵ.x, ϵ.y, ϵ.z), polarization direction, requires norm of 1

k::NamedTuple: (k.x, k.y, k.z), propagation direction, requires norm of 1

ϕ::Union{Function,Real}: time-dependent phase. of course, this can also be used to model a time-dependent detuning. Units are in radians. Note: if this is set to a function of time, then when constructing a Hamiltonian with the hamiltonian function, the units of time will be as specified by the timescale keyword argument.

pointing: an array of Tuple{Int,Real} for describing ion-laser pointing configuration. (first element of the tuple is the index for an ion and the second element is the scaling factor for the laser's Efield which must be between 0 and 1).

IonSim.LinearChain — Type

LinearChain(;
        ions, comfrequencies::NamedTuple{(:x,:y,:z)},
        selectedmodes::NamedTuple{(:x,:y,:z), N::Int=10
    )

Generates and stores all of the information necessary to describe a collection of ions trapped in a 3D harmonic potential and forming a linear coulomb crystal. Relevant to calculating the normal mode structure of the linear chain is the charge, mass and ordering of the ions.

user-defined fields

ions: An iterable list of ions that compose the linear Coulomb crystal

comfrequencies::NamedTuple{(:x,:y,:z): Describes the COM frequencies (x=ν_x, y=ν_y, z=ν_z). The z-axis is taken to be along the crystal's axis of symmetry. Note that these only correspond to true COM modes if all ions have the same mass, otherwise this is a misnomer and corresponds to the smallest (largest) eigenfrequency mode in the axial (radial) directions. Note: we assume that ν_x > ν_y > ν_z.

selectedmodes::NamedTuple{(:x,:y,:z)}: e.g. selectedmodes=(x=[1], y=[1, 2:3], z=[:]). Specifies the axis and a list of integers which correspond to the $i^{th}$ farthest mode away from the COM (see note above about meaning of "COM") for that axis. For example, selectedmodes=(z=[2]) would specify the axial stretch mode. These are the modes that will be modeled in the chain.Note: selectedmodes=(x=[],y=[],z=[1]), selectedmodes=(y=[],z=[1]) and selectedmodes=(;z=[1]) are all acceptable and equivalent.

N::Int=10: Optionally specify the Hilbert space dimension for each normal mode.

derived fields

ionpositions::Vector{<:Real}: The relative positions of the ions in meters.

IonSim.VibrationalMode — Type

VibrationalMode(
        ν::Real, modestructure::Vector{Real}, δν::Union{Function,Real}=0.; N::Int=10,
        axis::NamedTuple{(:x,:y,:z)}=ẑ
    )

user-defined fields

ν::Real: frequency [Hz]

modestructure::Vector{Real}: The normalized eigenvector describing the collective motion of the ions belonging to this mode.

δν::Union{Function,Real}: Either a function describing time-dependent fluctuations of ν or a real number which will be converted to the constant function t -> δν.

N::Int: Highest level included in the Hilbert space.

axis::NamedTuple{(:x,:y,:z)}: The axis of symmetry for the vibration. This must lie along one of the basis vectors x̂, ŷ or ẑ.

derived fields

shape::Vector{Int}: Indicates dimension of used Hilbert space (=[N+1]).

_cnst_δν::Bool: A Boolean flag signifying whether or not δν is a constant function.

Note: the iᵗʰ Fock state (|i⟩) can be obtained by indexing as v=VibrationalMode(...); v[i]

IonSim.LinearChain_fromyaml — Method

LinearChain_fromyaml(ions::Vector{<:Ion}, yaml::String; N::Int=10)

Load normal mode structure from the specified yaml file. It's up to the user to enforce physicality.

x:
  - frequency: 1e6
    mode: [0.1, 0.5, 0.3, 0.8]
  - frequency: 2e6
    mode: [0.3, 0.6, 0.5, 3]
y:
  - frequency: 8e6
    mode: [1, 1, 1, 1]
ionpositions: [-1, -0.5, -0.25, 5]

It is up to the user to enforce the physicality of this structure.

IonSim.axis — Method

axis(mode::VibrationalMode)

Returns mode.axis

IonSim.basis — Method

basis(chamber::Chamber)

Returns the composite basis describing the Hilbert space for chamber. This is the same as basis(iontrap(chain)).

IonSim.basis — Method

basis(chain::LinearChain)::CompositeBasis

Returns the composite basis describing the Hilbert space for chain.

Order is $ion₁ ⊗ ion₂ ⊗ ... ⊗ ion_N ⊗ mode₁ ⊗ mode₂ ⊗ ... ⊗ mode_N$, where the ion bases are ordered according to the order in ions(chain) and the vibrational modes are ordered according to the order in [xmodes(chain), ymodes(chain), zmodes(chain)].

IonSim.bfield! — Method

bfield!(chamber::Chamber, B::Real)

Sets chamber.B to B

IonSim.bfield — Method

bfield(chamber::Chamber, ion)

Retuns the value of the magnetic field in T at the location of ion, including both the trap's overall B-field and its B-field gradient. ion may be either an Ion or an Int indicating the desired ion's index within chamber.

IonSim.bfield — Method

bfield(chamber::Chamber)

Returns the magnitude of the (average) magnetic field in chamber, chamber.B,in T.

IonSim.bfield_fluctuation! — Method

bfield_fluctuation!(chamber::Chamber, δB::Function)

Sets chamber.δB to δB

IonSim.bfield_fluctuation! — Method

bfield_fluctuation!(chamber::Chamber, δB::Real)

Sets chamber.δB to a constant function t -> δB

IonSim.bfield_fluctuation — Method

bfield_fluctuation(chamber::Chamber)::Function

Returns the small magnetic field fluctuation chamber.δB in T as a function of time in s.

IonSim.bfield_unitvector! — Method

bfield_unitvector!(chamber::Chamber, Bhat::NamedTuple)

Sets chamber.Bhat to Bhat

IonSim.bfield_unitvector — Method

bfield_unitvector(chamber::Chamber)::NamedTuple{:x, :y, :z}

Returns the direction of the magnetic field inside chamber, chamber.Bhat

IonSim.bgradient! — Method

bgradient!(chamber::Chamber, ∇B::Real)

Sets chamber.∇B to ∇B

IonSim.bgradient! — Method

bgradient!(
        T::Chamber, ion_indxs::Tuple{Int,Int}, transition::Tuple, d::Real
    )

Sets the Bfield gradient in place to achieve a detuning d between the transition of two ions, which are assumed to be of the same species. ion_indxs refer to the ordering of the ions in the chain.

IonSim.bgradient — Method

bgradient(chamber::Chamber)

Returns the gradient of the magnetic field inside chamber, chamber.∇B, in T/m.

IonSim.charge — Method

charge(ion::Ion)

Returns the charge of ionin C.

IonSim.clearsublevelalias! — Method

clearsublevelalias!(I::Ion, sublevel)

Erases the assignment of an alias to sublevel of Ion I. Accepts either the full sublevel Tuple{String,Real} or its alias String. Also accepts a vector of sublevels to clear multiple alias assignments in a single call.

IonSim.clearsublevelalias! — Method

clearsublevelalias!(I::Ion)

Erases the assignment of all sublevel aliases of Ion I.

IonSim.coherentstate — Method

coherentstate(v::VibrationalMode, α::Number)

Returns a coherent state on v with complex amplitude $α$.

IonSim.comfrequencies — Method

comfrequencies(chain::LinearChain)

Returns chain.comfrequencies

IonSim.create — Method

create(v::VibrationalMode)

returns the creation operator for v such that: create(v) * v[i] = √(i+1) * v[i+1].

IonSim.destroy — Method

destroy(v::VibrationalMode)

Returns the destruction operator for v such that: destroy(v) * v[i] = √i * v[i-1].

IonSim.detuning! — Method

wavelength!(laser::Laser, Δ::Real)

Sets the detuning of laser to Δ.

IonSim.detuning — Method

detuning(laser::Laser)::Real

Returns the laser's detuning laser.Δ (in Hz)

IonSim.efield — Method

efield(laser::Laser)::Function

Returns the electric field amplitude (in V/m) of laser as a function of time.

IonSim.efield — Method

efield(I::Real)::Real

Returns the electric field (in V/m) corresponding to a light intensity of I (in W/m²) E = √(2I/(cϵ₀))

IonSim.einsteinA — Method

einsteinA(I::Ion, leveltransition::Tuple)

Returns Einstein A coefficient corresponding to the transition leveltransition[1] -> leveltransition[2]. The first level must be the lower level and the second must be the upper level.

IonSim.energy — Method

energy(I::Ion, level::String)

Returns the energy of level of I.

IonSim.energy — Method

energy(I::Ion, sublevel; B=0, ignore_manualshift=false)

Returns energy of sublevel of I. A Zeeman shift may be included by setting the value of the magnetic field B. The manual shift may be omitted by setting ignore_manualshift=true.

IonSim.frequency! — Method

frequency!(mode::VibrationalMode, ν::Real)

Sets mode.ν to ν

IonSim.frequency — Method

frequency(mode::VibrationalMode)

Returns mode.ν

IonSim.frequency_fluctuation! — Method

frequency_fluctuation!(mode::VibrationalMode, δν::Function)

Sets mode.δν to δν

IonSim.frequency_fluctuation! — Method

frequency_fluctuation!(mode::VibrationalMode, δν::Real)

Sets mode.δν to a constant function t -> δν

IonSim.frequency_fluctuation — Method

frequency_fluctuation(mode::VibrationalMode)

Returns mode.δν

IonSim.full_normal_mode_description — Method

full_normal_mode_description(ions, comfreqs::NamedTuple{(:x, :y, :z)})

Same thing but with an iterable list of ions and NamedTuple of COM frequencies explicitly given.

IonSim.full_normal_mode_description — Method

full_normal_mode_description(chain::LinearChain)

For each axis, this contains an array of tuples where the first element is a vibrational frequency [Hz] and the second element is a vector describing the participation of each ion at that vibrational frequency (i.e. the normal mode eigenvector corresponding to that eigenfrequency).

IonSim.full_normal_mode_description — Method

function full_normal_mode_description(
    M::Vector{<:Real}, Q::Vector{<:Int}, comfreqs::NamedTuple{(:x, :y, :z)}
)

Same thing but explicitly provide the masses M and charges Q of the ions.

IonSim.globalbeam! — Method

globalbeam!(laser, chamber::Chamber)

Set laser to shine with full intensity on all ions in Chamber. laser may be either a Laser or an Int indicating the desired laser's index within chamber.

IonSim.groundstate — Method

groundstate(obj)

If obj is a VibrationalMode, returns the N=0 ket of that mode. If obj is a Vector of VibrationalMode, returns a tensor product mode1[0] ⊗ mode2[0] ⊗ ... in the same order given. If obj is a LinearChain, returns the full ground state of the motional degrees of freedom as a tensor product.

IonSim.hamiltonian — Method

hamiltonian(
        chamber::Chamber; timescale::Real=1, lamb_dicke_order::Union{Vector{Int},Int}=1,
        rwa_cutoff::Real=Inf, displacement="truncated", time_dependent_eta=false
    )

Constructs the Hamiltonian for chamber as a function of time. Return type is a function h(t::Real, ψ) that, itself, returns a QuantumOptics.SparseOperator.

args

timescale: e.g. a value of 1e-6 will take time to be in μs

lamb_dicke_order: Only consider terms that change the phonon number by up to this value. If this is an Int, then the cutoff is applied to all modes. If this is a Vector{Int}, then lamb_dicke_order[i] is applied to the iᵗʰ mode, according to the order in basis(chamber). Note: this isn't quite the same thing as the Lamb-Dicke approximation since setting lamb_dicke_order=1 will retain, for example, terms proportional to $a^\dagger a$.

rwa_cutoff: drop terms in the Hamiltonian that oscillate faster than this cutoff.

displacement: This can be either "truncated"(default) or "analytic".

When an atom is irradiated, both the atom's energy and its momentum will generally be affected. For an atom in a harmonic potential, the exchange of momentum can be modeled as a displacement operation $D(α=iηe^{-iνt}) = exp[αa^† - α^*a]$, where $η$ is the Lamb-Dicke parameter, which can be described equivalently as either being proportional to the square root of the ratio of the recoil frequency with the ground state energy of the atom's motion or as the ratio of the spread of the ground state wavefunction to the wavelength of the laser.

When "truncated" is selected, the matrix elements of $D(α)$ are computed by constructing $α^* a, αa^†$ in a truncated basis (according to the dimension specified in your model) and then exponentiating their difference. This has the advantage, amongst other things, of guaranting unitarity.

If "analytic" is selected, then the matrix elements are computed assuming an infinite- dimensional Hilbert space.

For small displacements ($η ≪ N$, where $N$ is the dimension of the motion's Hilbert space), both of these methods will be good approximations.

time_dependent_eta::Bool: In addition to impacting the vibrational subspace directly, a change in the trap frequency, $δν$, will also change the Lamb-Dicke parameter. Since typically $δν≪ν$, this effect will be small $η ≈ η₀(1 + δν/2ν)$ and doesn't warrant the additional computational resources needed to calculate and update it in time. In this case, we can set time_dependent_eta=false (default), which will set $η(t) = η₀$.

IonSim.intensity! — Method

intensity(laser::Laser, I::Union{Function,Real})

Sets the intensity of laser to I.

IonSim.intensity — Method

intensity(laser::Laser)::Function

Returns the laser's intensity laser.I (in W/m²) as a function of time

IonSim.intensity_from_pitime! — Method

intensity_from_pitime!(
    laser::Laser, pi_time::Real, ion::Ion, transition::Tuple,
    Bhat::NamedTuple{(:x,:y,:z)}
)
intensity_from_pitime!(
    laser, pi_time::Real, ion, transition::Tuple, chamber::Chamber
)

Same as intensity_from_pitime, but updates laser[:I] in-place.

IonSim.intensity_from_pitime — Method

intensity_from_pitime(
    laser, pi_time::Real, ion, transition::Tuple, chamber::Chamber
    )

Compute the intensity needed to get a certain pi_time with a certain resonant laser-ion transition within chamber, which defines the magnetic field direction. laser may be either a Laser or an Int indicating the desired laser's index within chamber. ion may be either an Ion or an Int indicating the desired ion's index within chamber.

IonSim.intensity_from_pitime — Method

intensity_from_pitime(
    laser::Laser, pi_time::Real, ion::Ion, transition::Tuple,
    Bhat::NamedTuple{(:x,:y,:z)}
)

Compute the intensity needed to get a certain pi_time with a certain resonant laser-ion transition, in the presence of a magnetic field pointing in the direction Bhat.

IonSim.intensity_from_rabifrequency! — Method

intensity_from_rabifrequency!(
    laser::Laser, rabi_frequency::Real, ion::Ion, transition::Tuple,
    Bhat::NamedTuple{(:x,:y,:z)}
)
intensity_from_rabifrequency!(
    laser, rabi_frequency::Real, ion, transition::Tuple, chamber::Chamber
)

Same as intensity_from_rabifrequency!, but updates laser[:I] in-place.

IonSim.intensity_from_rabifrequency — Method

intensityfromrabifrequency( laser, rabifrequency::Real, ion, transition::Tuple, chamber::Chamber ) Compute the intensity needed to get a certain `rabifrequencywith a certain resonantlaser-iontransitionwithinchamber, which defines the magnetic field direction.lasermay be either a Laser or an Int indicating the desired laser's index withinchamber.ionmay be either an Ion or an Int indicating the desired ion's index withinchamber.laserandion` must either both be indices or both their respective Structs.

IonSim.intensity_from_rabifrequency — Method

intensity_from_rabifrequency(
    laser::Laser, rabi_frequency::Real, ion::Ion, transition::Tuple,
    Bhat::NamedTuple{(:x,:y,:z)}
)

Compute the intensity needed to get a certain rabi_frequency with a certain resonant laser-ion transition, in the presence of a magnetic field pointing in the direction Bhat.

IonSim.ionnumber — Method

ionnumber(ion::Ion)

Returns ion's number in its IonTrap ion.ionnumber If ion has not been added to an IonTrap, returns missing.

IonSim.ionposition — Method

ionposition(ion::Ion) Returns ion's position in its IonTrap, ion.ionposition in m. If ion has not been added to an IonTrap, returns missing.

IonSim.ionpositions — Method

ionpositions(chain::LinearChain)

Returns the positions of the ions in chain in meters.

IonSim.ionprojector — Method

ionprojector(obj, sublevels...; only_ions=false)

If obj<:IonTrap this will return $|ψ₁⟩⟨ψ₁|⊗...⊗|ψ\_N⟩⟨ψ\_N|⊗𝟙$ where $|ψᵢ⟩$ = obj.ions[i][sublevels[i]] and the identity operator $𝟙$ is over all of the COM modes considered in obj.

If only_ions=true, then the projector is defined only over the ion subspace.

If instead obj<:Chamber, then this is the same as obj = Chamber.iontrap.

IonSim.ions — Method

" ions(T::Chamber) Returns a list of the ions in the Chamber.

IonSim.ionstate — Method

ionstate(ion::Ion, sublevel)

Retuns the ket corresponding to the Ion being in state $|sublevel⟩$. Options: sublevel <: Tuple{String,Real}: Specifies full sublevel name sublevel <: String: Specifies sublevel alias sublevel <: Int: Returns the sublevelth eigenstate

IonSim.ionstate — Method

ionstate(object::Union{IonTrap, Chamber}, sublevels)

If N = length(ions(object)), returns N-dimensional ket corresponding to the ions being in the state $|sublevel₁⟩⊗|sublevel₂⟩⊗...⊗|sublevel\_N⟩$.

sublevels must be an length-N Vector, with each element specifying its corresponding ion's sublevel, using the same syntax as in ionstate(ion::Ion, sublevel).

IonSim.iontrap! — Method

iontrap!(chamber::Chamber, iontrap::IonTrap)

Sets chamber.iontrap to iontrap

IonSim.iontrap — Method

iontrap(chamber::Chamber)

Returns the IonTrap struct associated with chamber, chamber.iontrap

IonSim.lambdicke — Method

lambdicke(V::VibrationalMode, I::Ion, L::Laser)

The Lamb-Dicke parameter: $|k|cos(\theta)\sqrt{\frac{\hbar}{2m\nu}}$ for a given vibrational mode, ion and laser.

IonSim.landegf — Function

landegf(l::Real, j::Real, f::Real, i::Real, s::Real=1//2)

Landé g-factor of hyperfine energy level

args

l: orbital angular momentum quantum number

j: electron total angular momentum quantum number

f: total angular momentum quantum number

i: nuclear spin angular momentum quantum number

s: electronic spin angular momentum quantum number (defaults to 1/2)

IonSim.landegf — Method

landegf(I::Ion, level::String)

landegf for the quantum numbers of level in I.

IonSim.lasers! — Method

lasers!(chamber::Chamber, lasers::Vector{Laser})

Sets chamber.lasers to lasers

IonSim.lasers — Method

lasers(chamber::Chamber)::Vector{Laser}

Returns Vector of Lasers inside chamber, chamber.lasers

IonSim.level — Method

level(I::Ion, sublevel)

Retuns the energy level of I corresponding to sublevel.

IonSim.levels — Method

levels(I::Ion)

Returns array of all energy levels of I.

IonSim.leveltransitions — Method

leveltransitions(I::Ion)

Returns all allowed transitions between levels of I as a vector of Tuple{String,String}.

IonSim.lifetime — Method

lifetime(I::Ion, level::String)

Computes lifetime of level by summing the transition rates out of level. The sum is taken over all levels that the species may have, rather than the levels present in the instance I.

IonSim.manualshift! — Method

manualshift!(I::Ion, manualshift_dict::Dict)

Applies manualshift(I, sublevel, shift) to all pairs sublevel => shift of the Dict manual_shift_dict.

IonSim.manualshift! — Method

manualshift!(I::Ion, sublevel, shift::Real)

Applies a manual shift shift to the chosen sublevel of I (overwriting any previously assigned manual shift).

IonSim.manualshift — Method

manualshift(I::Ion, sublevel)

Returns the assigned manual shift of sublevel of Ion I.

IonSim.manualshift — Method

manualshift(ion::Ion)

Returns the Dict of manualshift for ion's energy levels ion.manualshift

IonSim.mass — Method

mass(ion::Ion)

Returns the mass of ion in kg.

IonSim.matrixelement — Function

matrixelement(ion::Ion, transition::Tuple, I::Real, ϵhat::NamedTuple, khat::NamedTuple, Bhat::NamedTuple=(;z=1))

Computes the matrix elements (units of Hz) between two energy sublevels args

ion: Ion undergoing transition

transition: Tuple of sublevels (full names or aliases) between which the transition is being calculated. Must be formatted such that energy(transition[2]) > energy(transition[1])

I: Intensity of the driving field

ϵhat: Unit vector of light polarization

khat: Unit vector of light wavevector

Bhat: Unit vector of magnetic field

IonSim.matrixelement — Method

matrixelement(ion, transition::Tuple, laser, chamber::Chamber, time::Real)

Calls matrixelement(ion, transition, I, ϵhat, khat, Bhat) with I, ϵhat, and khat evaluated for laser at time time, and Bhat evaluated for chamber.

ion may be either an Ion or an Int indicating the desired ion's index within chamber. laser may be either a Laser or an Int indicating the desired laser's index within chamber. ion and laser must either both be indices or both their respective Structs.

IonSim.modecutoff! — Method

modecutoff!(T::Chamber, N::Int)

Sets the upper bound of the Hilbert space of all VibrationalModes in the IonTrap of T to be the Fock state N.

IonSim.modecutoff! — Method

modecutoff!(lc::LinearChain, N::Int)

Sets the upper bound of the Hilbert space of all VibrationalModes in lc to be the Fock state N.

IonSim.modecutoff! — Method

modecutoff!(mode::VibrationalMode, N::Int)

Sets the upper bound of the Hilbert space of mode to be the Fock state N.

IonSim.modecutoff — Method

modecutoff

Returns mode.N

IonSim.modes — Method

modes(T::Chamber)

Returns modes(iontrap(T))

IonSim.modes — Method

modes(lc::LinearChain)

Returns an array of all of the selected VibrationalModes in the LinearChain. The order is [lc.x..., lc.y..., lc.z...].

IonSim.modestructure — Method

modestructure(mode::VibrationalMode)

Returns mode.modestructure

IonSim.nuclearspin — Method

nuclearspin(ion::Ion)

Returns the nuclear spin of ion.

IonSim.number — Method

number(v::VibrationalMode)

Returns the number operator for v such that: number(v) * v[i] = i * v[i].

IonSim.phase! — Method

phase!(laser::Laser, ϕ::Union{Function,Real})

Sets the phase of laser to ϕ.

IonSim.phase — Method

phase(laser::Laser)::Function

Returns the laser's phase laser.ϕ in (in radians) as a function of time

IonSim.pointing! — Method

pointing!(laser::Laser, p::Vector{Tuple{T1, T2}} where T1<:Int where T2<:Real)

Sets the pointing of laser with p. length(p) should be equal to the number of ions in the Chamber containing laser.

IonSim.pointing — Method

pointing(laser::Laser)::Vector

Return's the laser's pointing information laser.pointing if laser has been added to a Chamber.

IonSim.polarization! — Method

polarization!(laser::Laser, ϵ::ReNamedTuple{(:x, :y, :z)})

Sets the polarization of laser to ϵ.

IonSim.polarization — Method

polarization(laser::Laser)::NamedTuple{(:x, :y, :z)}

Returns the laser's polarization unit vector laser.ϵ

IonSim.quantumnumbers — Method

quantumnumbers(I::Ion, level::String)
quantumnumbers(I::Ion, sublevel)

Returns the quantum numbers of an energy level or sublevel of I as a NamedTuple. If second argument is a level, returns (:n, :i, :s, :l, :j, :f) If second argument is a sublevel, returns (:n, :i, :s, :l, :j, :f, :m)

IonSim.selectedmodes — Method

selectedmodes(chain::LinearChain)

Returns chain.selectedmodes

IonSim.shape — Method

shape(ion::Ion)

Returns the dimension of the ion's Hilbert space ion.shape

IonSim.shape — Method

shape(mode::VibrationalMode)

Returns mode.shape

IonSim.sigma — Method

sigma(ion::Ion, ψ1::sublevel[, ψ2::sublevel])

Returns $|ψ1\rangle\langle ψ2|$, where $|ψ_i\rangle$ corresponds to the state returned by ion[ψᵢ].

If ψ2 is not given, then $|ψ1\rangle\langle ψ1|$ is returned.

IonSim.speciesproperties — Method

speciesproperties(ion::Ion)

Returns the IonProperties struct of the ion species, which contains species-specific information including mass, energy levels, and transitions.

IonSim.sublevel — Method

sublevel(I::Ion, alias::String)

Returns the sublevel corresponding to the given alias alias of I. Inverse function of sublevelalias.

IonSim.sublevelalias! — Method

sublevelalias!(I::Ion, aliasassignments)

aliasassignments specifies which aliases should be assigned to which sublevels. There are two options to do this:

aliasassignments is a Vector of Tuples, with the first element of each being the sublevel (sublevel::Tuple{String,Real}) and the second being its assigned alias (alias::String)

aliasassignments is a Dict with the format alias::String => sublevel::Tuple{String,Real}

Calls sublevelalias!(I, sublevel, alias) for each pair sublevel, alias.

IonSim.sublevelalias! — Method

sublevelalias!(I::Ion, sublevel::Tuple{String,Real}, alias::String)

Assigns an alias alias to sublevel of I. This then allows one to pass alias in place of sublevel (for I only) into any function which accepts a sublevel as an argument.

IonSim.sublevelalias — Method

sublevelalias(I::Ion, sublevel::Tuple{String,Real})

Returns the alias assined to sublevel of I. If no alias is assigned, returns nothing.

IonSim.sublevelaliases — Method

sublevelaliases(ion::Ion)

Returns a Dict specifying all aliases assigned to sublevels of ion, in the format alias => sublevel, ion.sublevelalias

IonSim.sublevels — Method

sublevels(ion::Ion)

Returns the energy sublevels in the Hilbert space of ion, ion.sublevels

IonSim.subleveltransitions — Method

subleveltransitions(I::Ion)

Returns all allowed transitions between sublevels of I as a vector of Tuple{S,S} where S=Tuple{String,Real}.

IonSim.transitionfrequency — Method

transitionfrequency(ion, transition::Tuple, chamber::Chamber; ignore_manualshift=false)

Returns The frequency of the transition transition including the Zeeman shift experienced by ion given its position in T. ion may be either an Ion or an Int indicating the desired ion's index within chamber.

IonSim.transitionfrequency — Method

transitionfrequency(I::Ion, transition::Tuple; B=0, ignore_manualshift=false)

transition is a Tuple of two sublevels or levels.

Computes the absolute values of the difference in energies between transition[1] and transition[2].

If between sublevels, then the Zeeman shift may be included by setting the value of the magnetic field B, and manual shifts may be omitted by setting ignore_manualshift=true.

IonSim.transitionmultipole — Method

transitionmultipole(I::Ion, leveltransition::Tuple)

Returns the transition multiple ('E1', 'E2', etc.) corresponding to the transition leveltransition[1] -> leveltransition[2]. The first level must be the lower level and the second must be the upper level.

IonSim.transitionwavelength — Method

transitionwavelength(ion, transition::Tuple, chamber::Chamber; ignore_manualshift=false)

Returns The wavelength of the transition transition including the Zeeman shift experienced by ion given its position in T. ion may be either an Ion or an Int indicating the desired ion's index within chamber.

IonSim.transitionwavelength — Method

transitionwavelength(I::Ion, transition::Tuple; B=0, ignore_manualshift=false)

Returns the wavelength corresponding to transitionfrequency(I::Ion, transition::Tuple; B=0, ignore_manualshift=false).

IonSim.visualize — Method

visualize(lc::LinearChain, axis::String, modes::Vector{<:Int}, format="bars")

Visualize the normal mode structure of a lc as a Plots.jl plot. axis can either be one of "x", "y", "z" or a NamedTuple{(:x,:y,:z)}. modes is a vector of indices selecting which modes should be included in the plot. Slicing is supported. format can be either "bars" or "circles."

Note that the indexing refers to the full normal mode description and not the subset selectedmodes(lc).

IonSim.visualize — Method

visualize(
    vm::Tuple{Float64, Vector{Float64}}, axis::NamedTuple{(:x, :y, :z)}; format="bars"
)

Same thing but input a normal mode description as a tuple with first element the eigenfrequency and second the eigenvector.

IonSim.visualize — Method

visualize(vm::VibrationalMode; format="bars")

Visualize the normal mode structure of a vm as a Plots.jl plot. format can be either bar or ions. format can be either "bars" or "circles."

IonSim.wavelength! — Method

wavelength!(laser::Laser, λ::Real)

Sets the wavelength of laser to λ.

IonSim.wavelength — Method

wavelength(laser::Laser)::Real

Returns the laser's wavelength laser.λ (in m)

IonSim.wavelength_from_transition! — Method

wavelength_from_transition!(laser::Laser, ion, transition::Tuple, chamber::Chamber)

Sets the wavelength of laser to the transition wavelength of transition in the ion ion, at the magnetic field seen by ion in chamber. ion may be either an Ion or an Int indicating the desired ion's index within chamber.

IonSim.wavelength_from_transition! — Method

wavelength_from_transition!(laser::Laser, ion::Ion, transition::Tuple, B::Real)

Sets the wavelength of laser to the transition wavelength of transition in the ion ion, at a magnetic field value given by B.

IonSim.wavevector! — Method

wavevector!(laser::Laser, k::ReNamedTuple{(:x, :y, :z)})

Sets the wavevector of laser to k.

IonSim.wavevector — Method

wavevector(laser::Laser)::NamedTuple{(:x, :y, :z)}

Returns the laser's wavevector unit vector laser.k

IonSim.xmodes — Method

xmodes(T::Chamber)

Returns an array of all of the selected VibrationalModes in the x-direction in the Chamber's IonConfiguration.

IonSim.xmodes — Method

xmodes(lc::LinearChain)

Returns an array of all of the selected VibrationalModes in the x-direction in the LinearChain.

IonSim.ymodes — Method

ymodes(T::Chamber)

Returns an array of all of the selected VibrationalModes in the y-direction in the Chamber's IonConfiguration.

IonSim.ymodes — Method

ymodes(lc::LinearChain)

Returns an array of all of the selected VibrationalModes in the y-direction in the LinearChain.

IonSim.zeemanshift — Method

zeemanshift(I::Ion, sublevel, B::Real)

Returns the Zeeman shift at a magnetic field of B of sublevel of I. If sublevel has a custom g-factor defined, then this is used. Otherwise, landegf is used to compute the Landé g-factor. Zeeman shift calculated as $ΔE = (μ_B/ħ) ⋅ g_f ⋅ B ⋅ m / 2π$

IonSim.zeemanshift — Method

zeemanshift(I, sublevel, T::Chamber)

Calls zeemanshift(I::Ion, sublevel, B::Real) with B evaluated for ion I in T. I may be either an Ion or an Int indicating the desired ion's index within T.

IonSim.zeromanualshift! — Method

zeromanualshift!(I::Ion)

Sets the manual shift of all sublevels of I to zero.

IonSim.zmodes — Method

zmodes(T::Chamber)

Returns an array of all of the selected VibrationalModes in the z-direction in the Chamber's IonConfiguration.

IonSim.zmodes — Method

zmodes(lc::LinearChain)

Returns an array of all of the selected VibrationalModes in the z-direction in the LinearChain.

QuantumOpticsBase.coherentthermalstate — Method

coherentthermalstate(v::VibrationalMode, n̄::Real, α::Number; method="truncated)

Returns a displaced thermal state for v, which is created by applying a displacement operation to a thermal state. The mean occupation of the thermal state is n̄ and α is the complex amplitude of the displacement.

method can be either "truncated" or "analytic" and this argument determines how the displacement operator is computed (see: displace) .

QuantumOpticsBase.displace — Method

displace(v::VibrationalMode, α::Number; method="truncated")

Returns the displacement operator $D(α)$ corresponding to v.

If method="truncated" (default), the matrix elements are computed according to $D(α) = exp[αa^† - α^*a]$ where $a$ and $a^†$ live in a truncated Hilbert space of dimension modecutoff(v)+1. Otherwise if method="analytic", the matrix elements are computed assuming an infinite-dimension Hilbert space. In general, this option will not return a unitary operator.

QuantumOpticsBase.fockstate — Method

fockstate(v::VibrationalMode, N::Int)

Returns the fockstate $|N⟩$ on v.

QuantumOpticsBase.thermalstate — Method

thermalstate(v::VibrationalMode, n̄::Real; method="truncated")

Returns a thermal density matrix with $⟨a^†a⟩ ≈ n̄$. Note: approximate because we are dealing with a finite dimensional Hilbert space that must be normalized.

method can be set to either "truncated" (default) or "analytic". In the former case, the thermal density matrix is generated according to the formula: $ρ_{th} = exp(-νa^†a/T) / Tr [exp(-νa^†a/T)]$. In the later case, the analytic formula, assuming an infinite-dimensional Hilbert space, is used: $[ρ_{th}]_{ij} = δ_{ij} \frac{nⁱ}{(n+1)^{i+1}}.$

IonSim.analytical.molmersorensen2ion — Method

molmersorensen2ion(tspan, Ω::Real, ν::Real, δ::Real, η::Real, n̄::Real)

ref <br> Assumes vibrational mode starts in a thermal state with: $\langle a^\dagger a\rangle = n̄$ and ions start in doubly ground state. Returns (ρgg, ρee), the population in the doubly ground and doubly excited state, respectively. $[Ω], [ν], [δ] = Hz$

IonSim.analytical.rabiflop — Method

rabiflop(tspan, Ω::Real, η::Real, n̄::Real; s::Int=0) <br>

Single ion rabi flop. Returns: $\sum_{n=0}^∞ p_n sin^2(\Omega_n t)$ <br> with $\Omega_n = Ωe^{-η^2/2}η^s\sqrt{\frac{n!}{(n+s)!}}L_{n}^{s}(η^2)$ <br> where $s$ is the order of the (blue) sideband that we are driving and $L_{n}^{s}$ is the associated Laguerre polynomial. ref

IonSim.PhysicalConstants.c — Constant

c = 2.99792458e8 m/s <br> (speed of light in vacuum)

IonSim.PhysicalConstants.c_rank1 — Constant

c_rank1: Matrix of spherical basis vectors (defined in e.g. Quantum dynamics of cold trapped ions with application to quantum computation, Appl. Phys. B 66, 181-190 (1998).

Useful for converting between coordinates of rank-1 spherical tensors and Cartesian coordinates

IonSim.PhysicalConstants.c_rank2 — Constant

c_rank2: Matrix of spherical basis rank-2 tensors (defined in e.g. Quantum dynamics of cold trapped ions with application to quantum computation, Appl. Phys. B 66, 181-190 (1998).

Useful for converting between coordinates of rank-2 spherical tensors and Cartesian coordinates

IonSim.PhysicalConstants.e — Constant

e = 1.60217662e-19 C <br> (charge of electron)

IonSim.PhysicalConstants.eye3 — Constant

eye3: Three-dimensional identity matrix

IonSim.PhysicalConstants.kB — Constant

kB = 1.38064852e-23 $m^2kg/(s^2K)$

IonSim.PhysicalConstants.ħ — Constant

ħ = 1.0545718e-34 m²kg/s <br> (Planck's constant / 2π)

IonSim.PhysicalConstants.α — Constant

α = e²/4πϵ₀ħc``

IonSim.PhysicalConstants.μB — Constant

μB = 9.27400994e-24 J⋅T⁻¹ <br> (Bohr Magneton)

IonSim.PhysicalConstants.ϵ₀ — Constant

ϵ₀ = 8.85418782e-12 $(s^4A^2) / (m^3 kg)$